+++ evalml/pipelines/pipeline_base.py
@@ -178,7 +178,7 @@ class PipelineBase:
             objectives (list): Non-empty list of objectives to score on
 
         Returns:
-            float, dict:  score, ordered dictionary of other objective scores
+            dict: ordered dictionary of objective scores
         """
         if not isinstance(X, pd.DataFrame):
             X = pd.DataFrame(X)
@@ -190,7 +190,7 @@ class PipelineBase:
         y_predicted = None
         y_predicted_proba = None
 
-        scores = []
+        scores = OrderedDict()
         for objective in objectives:
             if objective.score_needs_proba:
                 if y_predicted_proba is None:
@@ -202,15 +202,11 @@ class PipelineBase:
                 y_predictions = y_predicted
 
             if objective.uses_extra_columns:
-                scores.append(objective.score(y_predictions, y, X))
+                scores.update({objective.name: objective.score(y_predictions, y, X)})
             else:
-                scores.append(objective.score(y_predictions, y))
-        if not objectives:
-            return scores[0], {}
+                scores.update({objective.name: objective.score(y_predictions, y)})
 
-        other_scores = OrderedDict(zip([n.name for n in objectives[1:]], scores[1:]))
-
-        return scores[0], other_scores
+        return scores
 
     def get_plot_data(self, X, y, plot_metrics):
         """Generates plotting data for the pipeline for each specified plot metric
@@ -230,7 +226,7 @@ class PipelineBase:
             y = pd.Series(y)
         y_predicted = None
         y_predicted_proba = None
-        scores = []
+        scores = OrderedDict()
         for plot_metric in plot_metrics:
             if plot_metric.score_needs_proba:
                 if y_predicted_proba is None:
@@ -240,8 +236,7 @@ class PipelineBase:
                 if y_predicted is None:
                     y_predicted = self.predict(X)
                 y_predictions = y_predicted
-            scores.append(plot_metric.score(y_predictions, y))
-        scores = OrderedDict(zip([n.name for n in plot_metrics], scores))
+            scores.update({plot_metric.name: plot_metric.score(y_predictions, y)})
         return scores
 
     @property